% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{animalesSmall}
\alias{animalesSmall}
\title{Small animals dataset.}
\description{
A (cropped) dataset polling children perception of animals characteristics.
}
\author{
Xavier Valls\email{xaviervallspla@gmail.com}
}
\references{
""
}
\keyword{data}

