% Generated by roxygen2 (4.0.1): do not edit by hand
\name{callASIAlgorithm}
\alias{callASIAlgorithm}
\title{Functions that performs the algorithm for the Statistical implicative analysis.}
\usage{
callASIAlgorithm(data, model, report = FALSE)
}
\arguments{
\item{data}{External data.}

\item{model}{Probability model to use. Binom or pois.}

\item{report}{If true, displays a report}
}
\value{
report 		Report of the data: algorithm results, hierarchical tree, basic statistic analysis.

newickTree	Hierarchical tree in Newick's format.
}
\description{
Functions that performs the algorithm for the Statistical implicative analysis.
}
\author{
Xavier Valls \email{xaviervallspla@gmail.com}
}

