% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ASImodel}
\alias{ASImodel}
\title{A Reference Class to represent the Statistical Implicative Analysis model.}
\description{
A Reference Class to represent the Statistical Implicative Analysis model.
}
\section{Fields}{

\describe{
\item{\code{data}}{External data.}

\item{\code{model}}{Probability model to use. Binom or pois.}

\item{\code{joinMatrix}}{Matrix of joint classes}

\item{\code{similarityMatrix}}{Matrix of bivariant similarities.}

\item{\code{genericImplicationsMatrix}}{Matrix of generic Implications}

\item{\code{nPrimitiveClasses}}{Number of primitive classes}
}}

