% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{ASImodel}
\alias{ASImodel}
\title{A Reference Class to represent the Statistical Implicative Analysis model.}
\description{
A Reference Class to represent the Statistical Implicative Analysis model.
}
\section{Fields}{

\describe{
\item{\code{data}}{External data.}

\item{\code{model}}{Probability model to use. Binom or pois.}

\item{\code{joinMatrix}}{Matrix of joint classes}

\item{\code{similarityMatrix}}{Matrix of bivariant similarities.}

\item{\code{genericImplicationsMatrix}}{Matrix of generic Implications}

\item{\code{nPrimitiveClasses}}{Number of primitive classes}
}}
\section{Methods}{

\describe{
\item{\code{getMaximumSimilarity(at.level = levels)}}{Gets the maximum similartity at the specified level. By default the last one}

\item{\code{getSignificativeNodes()}}{Gets the significative nodes of the hierarchical tree}

\item{\code{getSimilarityMatrix(at.level = levels)}}{Returns the similarity matrix at the specified level of the hierarchical tree}

\item{\code{joinClasses(Tuple)}}{Joins two classes into a new node of the hierarchical tree}

\item{\code{setContribution(genericPair, at.level = levels)}}{Computes the contribution of each individual for a certain class}

\item{\code{setTypicality(genericPair, at.level = levels)}}{Computes the typicality of each individual for a certain class}
}}

